octeontx: mbox: prevent erroneous tag mismatch by casting addition result to u16
Even though mbox->tag_own is u16, mbox->tag_own + 1 is promoted to
int and this causes mismatch when mbox->tag_own = 0xffff and it gets
compared to 0 (ram_hdr.tag). Solved by casting the result to u16.

[Kevin: The original patch got from Marvell sdk10.0_19.06]

+++ drivers/net/ethernet/cavium/octeontx-83xx/octeontx_mbox.c
@@ -127,7 +127,7 @@ int mbox_send(struct mbox *mbox, struct mbox_hdr *hdr, const void *txmsg,
 	}
 	if (!wait)
 		return -1; /* timeout */
-	if ((mbox->tag_own + 1) != ram_hdr.tag)
+	if ((u16)(mbox->tag_own + 1) != ram_hdr.tag)
 		return -1; /* tag mismatch */
 	(mbox->tag_own)++;
